PERFORCE_INTEGRATION_ENABLED switch can be used to disable P4 communications

+++ projectaccess/group_actions.py
@@ -1,4 +1,6 @@
 
+from django.conf import settings
+
 from P4Connection import P4ConnectionAsServiceUser
 from models import PAUser, PAGroup
 
@@ -14,14 +16,15 @@ def create_new_group(group_name):
 
 def delete_group(group):
 
-    with P4ConnectionAsServiceUser() as p4:
+    if settings.PERFORCE_INTEGRATION_ENABLED:
+        with P4ConnectionAsServiceUser() as p4:
 
-        # Delete group in P4
-        # TODO: perhaps check for group existence before deleting? That will allow for ignoring a narrower range of exceptions
-        try:
-            p4.delete_group(group.name)
-        except:
-            pass
+            # Delete group in P4
+            # TODO: perhaps check for group existence before deleting? That will allow for ignoring a narrower range of exceptions
+            try:
+                    p4.delete_group(group.name)
+            except:
+                pass
 
     group.delete()
 
@@ -29,16 +32,17 @@ def delete_group(group):
 
 def add_user_to_group(group, user):
 
-    with P4ConnectionAsServiceUser() as p4:
+    group.members.add(user)
+    group.save()
 
-        group.members.add(user)
-        group.save()
-        p4.add_user_to_group(str(group.name), str(user.p4_user_name))
+    if settings.PERFORCE_INTEGRATION_ENABLED:
+        with P4ConnectionAsServiceUser() as p4:
+            p4.add_user_to_group(str(group.name), str(user.p4_user_name))
 
 def remove_user_from_group(group, user):
 
-    with P4ConnectionAsServiceUser() as p4:
-
-        group.members.remove(user)
-        group.save()
-        p4.remove_user_from_group(str(group.name), str(user.p4_user_name))
+    group.members.remove(user)
+    group.save()
+    if settings.PERFORCE_INTEGRATION_ENABLED:
+        with P4ConnectionAsServiceUser() as p4:
+            p4.remove_user_from_group(str(group.name), str(user.p4_user_name))
